refactor(scope): blueprints are used in flask The code was refactored to make the backend modular with flask. The routes /api/email were created to send emails to the users who have registered to the service and /api/sheet to store the registration data
