+++ server/api/email/routes.py
@@ -1,0 +1,8 @@
+from flask import jsonify
+
+from . import route
+
+
+@route.route("/")
+def index():
+    return jsonify({"email": "index"})
